# Stop encoding in secure_compare for speedup