Hex encoding is base 16 which makes the original input twice as big. We don't need to encode the SHA256 output bytes and the shorter output means we spend less time `fixed_length_secure_compare`.

Benchmark script:
```ruby
# frozen_string_literal: true

require "bundler/inline"

gemfile(true) do
  source "https://rubygems.org"

  git_source(:github) { |repo| "https://github.com/#{repo}.git" }

  gem "rails", github: "rails/rails"
  gem "benchmark-ips"
end

require "active_support"

module ActiveSupport
  module SecurityUtils
    def fast_secure_compare(a, b)
      fixed_length_secure_compare(::Digest::SHA256.digest(a), ::Digest::SHA256.digest(b)) && a == b
    end
    module_function :fast_secure_compare
  end
end

SCENARIOS = {
  "Same strings, length 16" => ["a" * 16, "a" * 16],
  "Different strings, length 16" => ["a" * 16, "b" * 16],
  "Same strings, length 32" => ["a" * 32, "a" * 32],
  "Different strings, length 32" => ["a" * 32, "b" * 32]
}

SCENARIOS.each_pair do |name, (a, b)|
  puts
  puts " #{name} ".center(80, "=")
  puts

  Benchmark.ips do |x|
    x.report("secure_compare") { ActiveSupport::SecurityUtils.secure_compare(a, b) }
    x.report("fast_secure_compare") { ActiveSupport::SecurityUtils.fast_secure_compare(a, b) }
    x.compare!
  end
end
```

Results from my MacBook 2018 (2.2 GHz Intel Core i7):
```
=========================== Same strings, length 16 ============================
Comparison:
 fast_secure_compare:   157901.1 i/s
      secure_compare:    91265.1 i/s - 1.73x  slower

========================= Different strings, length 16 =========================
Comparison:
 fast_secure_compare:   158115.6 i/s
      secure_compare:    91069.1 i/s - 1.74x  slower

=========================== Same strings, length 32 ============================
Comparison:
 fast_secure_compare:   158770.5 i/s
      secure_compare:    91677.7 i/s - 1.73x  slower

========================= Different strings, length 32 =========================
Comparison:
 fast_secure_compare:   156523.3 i/s
      secure_compare:    92266.6 i/s - 1.70x  slower
```