@kaspth 2019-03-09
LGTM! Let's just get the eyes of one other @rails/security member first.