@rafaelfranca 2019-03-26
@matthewd @tenderlove I asked your review because I know you were discussing this part of the code in the last month.