@bdewater 2019-03-26
Looking at this again I realized base64 encoding (1.25x difference compared to current) is not needed at all, we can just use the SHA256 output bytes as-is (1.73x difference) 🏎