@matthewd 2019-03-27
> we can just use the SHA256 output bytes as-is

👌🏻

(though I think I'd be concerned if this was a hot path for someone)